requirements.txt
openai
requests
Pillow
transformers
langdetect
pytesseract
tesseract
